langchain-ai-agent/requirements.txt
duckduckgo-search
langchain 
langchain-community 
langchain-anthropic
langchain-openai
pydantic
python-dotenv
wikipedia 